requirements.txt
aiogram==3.17.0
celery==5.4.0
Django==5.1.5
django-celery-beat==2.7.0
djangorestframework==3.15.2
pillow==11.1.0
whitenoise==6.8.2
redis==5.2.0
gunicorn==21.2.0
pytz==2025.1
django-cors-headers==4.3.1
